requirements.txt
nibabel
matplotlib==3.1.3
SimpleITK>=2.2.0
itkwidgets
itk-elastix
vtk
scikit-image
scikit-learn
monai
torch
pandas
wandb
pynrrd